Crate osproto

Protocol Structures for OpenStack API

Introduction

This crate is a collection of structures written by hand based on OpenStack API reference documentation, adapting it for more native Rust look and feel.

This crate does not contain any code to access OpenStack API. For low-level asynchronous SDK, check out rust-osauth, for a more high-level API see rust-openstack.

Stability

This crate is unstable by design. Particularly, two kinds of breaking changes will be regularly made:

Adding new public fields to existing structures.

Making required fields optional.

Downstream crates as supposed to lock a single version of the crate and not expose its structures as part of their public API.

Modules

common
Common JSON structures and protocol bits.

identity
Identity V3 JSON structures and protocol bits.