Module osproto::common

Common JSON structures and protocol bits.

Structs

IdAndName
A reference to an ID and name.

IntoStableIter
An iterator over stable versions.

Link
A link to a resource.

Ref

Version
A single API version as returned by a version discovery endpoint.

XdotY
A pair X.Y where X and Y can be converted to/from a string and Y is optional.

Enums

Root
A result of a version discovery endpoint.

Functions

empty_as_default
Deserialize a value where empty string is replaced by Default value.